Refactor completed - beat manager and video player state logic has been moved to viewmodel ownership and abstracted from UI

+++ app/src/main/java/com/example/video_playbacker_android/network/responses.kt
@@ -66,5 +66,5 @@ data class BeatsResponse(
     val bpm: Float,
 
     @SerialName("beat_frames")
-    val beatFrames: List<Int>
+    val beatFrames: List<Float>
 )
